requirements.txt
darkdetect
fcmeans
h5py
matplotlib
natsort
numba
numpy
open_clip_torch
pandas
Pillow
PyQt5
pyqtgraph
scikit-learn
scipy
torch
torchvision
umap-learn